refactor(tests): move CellRef tests aside
